Move Feedback tab and page to Settings (#1072)
* Remove Feedback tab from ShellPage and add to Settings

* Remove unused usings, change resources and icon

* Remove system management dependency

---------
