+++ settings/DevHome.Settings/NativeMethods.txt
@@ -1,0 +1,4 @@
+﻿GetPhysicallyInstalledSystemMemory
+GlobalMemoryStatusEx
+GetSystemInfo
+CoCreateInstance
