+++ src/App.xaml.cs
@@ -87,8 +87,6 @@ public partial class App : Application, IApp
             services.AddSingleton<WindowEx>(_ => MainWindow);
 
             // Views and ViewModels
-            services.AddTransient<FeedbackViewModel>();
-            services.AddTransient<FeedbackPage>();
             services.AddTransient<ShellPage>();
             services.AddTransient<InitializationPage>();
             services.AddTransient<ShellViewModel>();
